+++ sw.js
@@ -118,6 +118,7 @@ const SHELL_ASSETS = [
   './src/study-tracker.js',
   './src/spaced-repetition.js',
   './src/recommendations.js',
+  './src/command-palette.js',
   './src/views/dictionary.js',
   './content/exams/cosmetic/교재/understanding/images/피부의 구조 단면도 보완_인포그래픽.png',
   './content/exams/cosmetic/교재/understanding/images/모발의 구조 단면도 보완_인포그래픽.png',